Task #10285 As an Developer i want to check user http session Task #10300 As an developer i want to replace currency from $ to views

+++ iphone/pipture/Pipture/Classes/PiptureAppDelegate.h
@@ -33,7 +33,7 @@
 
 @interface PiptureAppDelegate : UIResponder <UIApplicationDelegate,DataRequestProgress, AuthenticationDelegate, BalanceReceiver, UINavigationControllerDelegate, UIAlertViewDelegate>
 {
-    float balance;
+    int balance;
     InAppPurchaseManager * purchases;
     NetworkConnection curConnection;
     NetworkConnectionInformer * wifiConnection;
